+++ config/module.config.php
@@ -78,6 +78,11 @@ return [
          *   boolean false, the value of this flag will define whether or
          *   not additional fields present in the payload will be merged
          *   with the filtered data.
+         *
+         * - remove_empty_data: Should we remove empty data from received data?
+         *   - If no `remove_empty_data` flag is present, do nothing - use data as is
+         *   - If `remove_empty_data` flag is present AND is boolean true, then remove
+         *     empty data from current data array
          */
     ],
 ];